added a prime multiplication solution.
Paradoxically, this solution should benifit the most from chaching and branch prediction.
There are no branches, and the only random lookups are in a rainbow table of primes, which in theory should get cached quickly.
However, The product variable RAPIDLY grows to outscale hardware optimizations.  Multiplying 100000000 digit numbers is NOT fast at all.
This is essentially a variant of the numeric addition solution, but with multiplication.  Due to the reasons listed above, this algorithm performs almost exclusivly worse than addition.